publish_date_modification2

+++ Modules/QuestionBank/Http/Controllers/QuizLessonController.php
@@ -124,6 +124,10 @@ class QuizLessonController extends Controller
         $quizLesson = QuizLesson::where('quiz_id',$request->quiz_id)
                         ->where('lesson_id',$request->quiz_id)->first();
 
+        if(!isset($quizLesson)){
+            return HelperController::api_response_format(404, null,'This quiz doesn\'t belongs to the lesson');
+        }
+
         $quizLesson->update([
             'quiz_id' => $request->quiz_id,
             'lesson_id' => $request->lesson_id,
@@ -153,6 +157,10 @@ class QuizLessonController extends Controller
 
         $quizLesson = QuizLesson::where('quiz_id',$request->quiz_id)
                 ->where('lesson_id',$request->quiz_id)->first();
+
+        if(!isset($quizLesson)){
+            return HelperController::api_response_format(404, null,'This quiz doesn\'t belongs to the lesson');
+        }
 
         $quizLesson->delete();
